requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
tqdm
spacy
sentence-transformers
faiss-cpu
google-genai
python-dotenv
requests
numpy